broke out into multiple files and renames a couple functions
